+++ include/manager/TMapObjBaseManager.h
@@ -3,6 +3,7 @@
 
 #include "JDrama.h"
 #include "JSUStream.h"
+#include "types.h"
 #include "tmapobj/TMapObjBase.h"
 #include "manager/TLiveManager.h"
 
@@ -17,14 +18,14 @@ class TMapObjBaseManager : public TLiveManager
 	void load(JSUMemoryInputStream &);
 	void createModelData();
 	void clipActors(TGraphics *);
-	TMapObjBase* makeObjAppeared(unsigned long);
-	TMapObjBase* makeObjAppear(unsigned long);
-	TMapObjBase* makeObjAppear(float, float, float, unsigned long, bool);
+	TMapObjBase* makeObjAppeared(u32);
+	TMapObjBase* makeObjAppear(u32);
+	TMapObjBase* makeObjAppear(f32, f32, f32, u32, bool);
 
-	static TMapObjBase* newAndRegisterObjByEventID(unsigned long, char const *);
+	static TMapObjBase* newAndRegisterObjByEventID(u32, char const *);
 	
-	float _38; // clip actor related
-	float _3C; // ^^
+	f32 _38; // clip actor related
+	f32 _3C; // ^^
 };
 
 #endif // TMAPOBJBASEMANAGER_H
