+++ include/spc/TSpcInterp.h
@@ -1,19 +1,20 @@
 #ifndef TSPCINTERP_H
 #define TSPCINTERP_H
 
+#include "types.h"
 #include "spc/TSpcBinary.h"
 
 class TSpcInterp
 {
     public:
-    TSpcInterp(TSpcBinary *, void *, int, int, int, int);
+    TSpcInterp(TSpcBinary *, void *, u32, u32, u32, u32);
     ~TSpcInterp();
 
     void update();
-    void verifyArgNum(unsigned long, unsigned long *);
+    void verifyArgNum(u32, u32 *);
     void dump();
-    void dispatchBuiltin(unsigned long, unsigned long);
-    void dispatchBuiltinDefault(unsigned long, unsigned long);
+    void dispatchBuiltin(u32, u32);
+    void dispatchBuiltinDefault(u32, u32);
 
     /* SPC EXECUTE FUNCTIONS */
     void execend();
@@ -25,29 +26,29 @@ class TSpcInterp
     void execret0();
 
     TSpcBinary* spcBin; // _0
-    int _4;
-    int _8;
-    int _C;
-    int _10;
-    int _14;
-    int _18;
-    int _1C;
-    int _20;
-    int _24;
-    int _28;
-    int _2C;
-    int _30;
-    int _34;
-    int _38;
-    int _3C;
-    int _40;
-    int _44;
-    int _48;
-    int _4C;
-    int _50;
-    int _54;
-    int _58;
-    int _5C;
+    u32 _4;
+    u32 _8;
+    u32 _C;
+    u32 _10;
+    u32 _14;
+    u32 _18;
+    u32 _1C;
+    u32 _20;
+    u32 _24;
+    u32 _28;
+    u32 _2C;
+    u32 _30;
+    u32 _34;
+    u32 _38;
+    u32 _3C;
+    u32 _40;
+    u32 _44;
+    u32 _48;
+    u32 _4C;
+    u32 _50;
+    u32 _54;
+    u32 _58;
+    u32 _5C;
 };
 
 #endif // TSPCINTERP_H
