+++ include/spc/TSpcTypedInterp.h
@@ -1,6 +1,7 @@
 #ifndef TSPCTYPEDINTERP_H
 #define TSPCTYPEDINTERP_H
 
+#include "types.h"
 #include "spc/TSpcInterp.h"
 
 template <typename T>
@@ -9,7 +10,7 @@ class TSpcTypedInterp : public TSpcInterp
     public:
     ~TSpcTypedInterp();
 
-    void dispatchBuiltin(unsigned long, unsigned long);
+    void dispatchBuiltin(u32, u32);
 };
 
 #endif // TSPCTYPEDINTERP_H
