+++ include/tmapobj/TMapObjSwitch.h
@@ -12,17 +12,17 @@ class TMapObjSwitch : public TMapObjBase
     ~TMapObjSwitch();
 
     void load(JSUMemoryInputStream &);
-    bool recieveMessage(THitActor * hitActor, unsigned long);
+    bool recieveMessage(THitActor * hitActor, u32);
     void control();
     
-    int _138;
-    int _13C;
-    int _140;
-    int _144;
-    short _148;
-    short _14A;
-    short _14C;
-    short _14E;
+    u32 _138;
+    u32 _13C;
+    u32 _140;
+    u32 _144;
+    u16 _148;
+    u16 _14A;
+    u16 _14C;
+    u16 _14E;
 };
 
 #endif // TMAPOBJSWITCH_H
